requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
cassandra-driver==3.28.0
cassio==0.1.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distlib==0.3.7
distro==1.8.0
filelock==3.13.1
frozenlist==1.4.0
geomet==0.2.1.post1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.342
langchain-core==0.0.7
langsmith==0.0.67
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.6
packaging==23.2
platformdirs==4.0.0
pluggy==1.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
virtualenv==20.24.7
yarl==1.9.3
